Refactor self test controllers (PP-497) (#1630)
* Refactor self tests
* Add some additional comments
