+++ pyproject.toml
@@ -68,13 +68,11 @@ module = [
     "api.admin.announcement_list_validator",
     "api.admin.config",
     "api.admin.controller.catalog_services",
-    "api.admin.controller.collection_self_tests",
     "api.admin.controller.collection_settings",
     "api.admin.controller.discovery_service_library_registrations",
     "api.admin.controller.discovery_services",
     "api.admin.controller.integration_settings",
     "api.admin.controller.library_settings",
-    "api.admin.controller.patron_auth_service_self_tests",
     "api.admin.controller.patron_auth_services",
     "api.admin.dashboard_stats",
     "api.admin.form_data",
